Move test package to pkg/testutil
