+++ storage/local/persistence_test.go
@@ -20,10 +20,10 @@ import (
 
 	clientmodel "github.com/prometheus/client_golang/model"
 
+	"github.com/prometheus/prometheus/pkg/testutil"
 	"github.com/prometheus/prometheus/storage/local/codable"
 	"github.com/prometheus/prometheus/storage/local/index"
 	"github.com/prometheus/prometheus/storage/metric"
-	"github.com/prometheus/prometheus/utility/test"
 )
 
 var (
@@ -34,16 +34,16 @@ var (
 	m5 = clientmodel.Metric{"label": "value5"}
 )
 
-func newTestPersistence(t *testing.T, encoding chunkEncoding) (*persistence, test.Closer) {
+func newTestPersistence(t *testing.T, encoding chunkEncoding) (*persistence, testutil.Closer) {
 	*defaultChunkEncoding = int(encoding)
-	dir := test.NewTemporaryDirectory("test_persistence", t)
+	dir := testutil.NewTemporaryDirectory("test_persistence", t)
 	p, err := newPersistence(dir.Path(), false, false, func() bool { return false })
 	if err != nil {
 		dir.Close()
 		t.Fatal(err)
 	}
 	go p.run()
-	return p, test.NewCallbackCloser(func() {
+	return p, testutil.NewCallbackCloser(func() {
 		p.close()
 		dir.Close()
 	})
